backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
pdfplumber
Pillow
pytesseract
python-docx
requests
groq
python-dotenv
pandas
numpy
sqlalchemy
aiofiles
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
google-search-results
psycopg2-binary

